truc truc

+++ helper.h
@@ -1,7 +1,7 @@
 #ifndef HELPER_H
 #define HELPER_H
 
-double benchmark(void (*function)(int*, int), int* tab, int nb, int printcsl);
+long benchmark(void (*function)(int*, int), int* tab, int nb, int printcsl);
 void afficheTableau(int* tab, int tabSize);
 void afficheTableau10(int* tab, int tabSize);
 int* randTab(int nb, int maxValue);
